Some minor changes in arch/screen.* and lib/lib.*

+++ lib/lib.h
@@ -3,13 +3,12 @@
 #define __LIB_H__
 
 #include "types.h"
-#include "kprintf.h"
 
 // Write a string at the cursor's current position
 int puts(const char* str);
 
 // Some memory functions
-void* memset(void* ptr, int value, size_t num);
+void* memset(void* ptr, uint8_t value, size_t num);
 void* memcpy(void* destination, const void* source, size_t num);
 
 void panic(const char *format, ...);
